requirements.txt
# pip install -r requirements.txt

opencv-python
pickle
pillow
scikit-learn
torch
torchvision
tensorflow